[CONTROLLER/HTTP] moves resource api
