+++ server/controller/http/appender/go.mod
@@ -1,0 +1,3 @@
+module github.com/deepflowio/deepflow/server/controller/http/appender
+
+go 1.18
